Move common dwarf things out of cu.h to dwarf.h
